Move utility function
