+++ src/jobs/measure-fills/measure-fill.js
@@ -4,7 +4,7 @@ const signale = require('signale');
 const { getToken } = require('../../tokens/token-cache');
 const formatTokenAmount = require('../../tokens/format-token-amount');
 const getConversionRate = require('../../rates/get-conversion-rate');
-const getMeasurableActor = require('./get-measurable-actor');
+const getMeasurableActor = require('../../fills/get-measurable-actor');
 const normalizeSymbol = require('../../tokens/normalize-symbol');
 
 const logger = signale.scope('measure fill');
